Implement simulate_taq_upload for quotes too.
