+++ bigtable/api/collate_taq.cc
@@ -27,10 +27,6 @@ namespace {
 // ... save ourselves some typing ...
 namespace bigtable = ::google::bigtable::v2;
 
-void collate_quotes(std::shared_ptr<grpc::ChannelCredentials> credentials,
-                    std::string const& table_prefix,
-                    std::string const& yyyymmdd, int report_progress_rate,
-                    int batch_size);
 void collate_trades(std::shared_ptr<grpc::ChannelCredentials> credentials,
                     std::string const& table_prefix,
                     std::string const& yyyymmdd, int report_progress_rate,
@@ -73,17 +69,12 @@ int main(int argc, char* argv[]) try {
   std::string table_prefix =
       "projects/" + project_id + "/instances/" + instance_id + "/tables/";
 
-  std::thread quotes([=]() {
-      collate_quotes(credentials, table_prefix, yyyymmdd, report_progress_rate,
-		     batch_size);
-    });
   std::thread trades([=]() {
       collate_trades(credentials, table_prefix, yyyymmdd, report_progress_rate,
 		     batch_size);
     });
 
   trades.join();
-  quotes.join();
 
   return 0;
 } catch (std::exception const& ex) {
@@ -123,120 +114,6 @@ void append_to_request(bigtable::MutateRowsRequest& request,
 // Perform a Bigtable::MutateRows() request until all mutations complete.
 void mutate_with_retries(bigtable::Bigtable::Stub& bt_stub,
                          bigtable::MutateRowsRequest& req);
-
-void collate_quotes(std::shared_ptr<grpc::ChannelCredentials> credentials,
-                    std::string const& table_prefix,
-                    std::string const& yyyymmdd, int report_progress_rate,
-                    int batch_size) {
-  // ... notice that Bigtable has separate endpoints for different APIs,
-  // we are going to upload some data, so the correct endpoint is:
-  auto channel = grpc::CreateChannel("bigtable.googleapis.com", credentials);
-
-  std::unique_ptr<bigtable::Bigtable::Stub> bt_stub(
-      bigtable::Bigtable::NewStub(channel));
-
-  // ... prepare the request to update the destination table ...
-  bigtable::MutateRowsRequest mutation;
-  mutation.set_table_name(table_prefix + "daily");
-  Quotes quotes;
-  using namespace std::chrono;
-  auto upload_start = steady_clock::now();
-  int count = 0;
-  int chunk_count = 0;
-
-  // ... once the table is created and the data uploaded we can make
-  // the query ...
-  bigtable::ReadRowsRequest request;
-  request.set_table_name(table_prefix + "raw-quotes");
-  auto& chain = *request.mutable_filter()->mutable_chain();
-  // ... first only accept the "taq" column family ...
-  chain.add_filters()->set_family_name_regex_filter("taq");
-  // ... then only accept the "quote" column ...
-  chain.add_filters()->set_column_qualifier_regex_filter("quote");
-
-  grpc::ClientContext context;
-  auto stream = bt_stub->ReadRows(&context, request);
-
-  std::string current_row_key;
-  std::string current_column_family;
-  std::string current_column;
-  std::string current_value;
-  // ... receive the streaming response ...
-  bigtable::ReadRowsResponse response;
-  while (stream->Read(&response)) {
-    ++chunk_count;
-    for (auto& cell_chunk : *response.mutable_chunks()) {
-      if (not cell_chunk.row_key().empty()) {
-        current_row_key = cell_chunk.row_key();
-      }
-      if (cell_chunk.has_family_name()) {
-        current_column_family = cell_chunk.family_name().value();
-        if (current_column_family != "taq") {
-          throw std::runtime_error(
-              "strange, only 'taq' family name expected in the query");
-        }
-      }
-      if (cell_chunk.has_qualifier()) {
-        current_column = cell_chunk.qualifier().value();
-        if (current_column != "quote") {
-          throw std::runtime_error(
-              "strange, only 'quote' column expected in the query");
-        }
-      }
-      if (cell_chunk.timestamp_micros() != 0) {
-        throw std::runtime_error(
-            "strange, only the 0 timestamp expected in the query");
-      }
-      if (cell_chunk.value_size() > 0) {
-        current_value.reserve(cell_chunk.value_size());
-      }
-      current_value.append(cell_chunk.value());
-      if (cell_chunk.commit_row()) {
-	++count;
-        // ... process this cell, we want to convert the value into a Quotes
-        // proto ...
-        Quote q;
-        if (not q.ParseFromString(current_value)) {
-          std::cerr << current_row_key << ": ParseFromString() failed"
-                    << std::endl;
-          continue;
-        }
-        if (quotes.ticker() != q.ticker()) {
-          if (not quotes.ticker().empty()) {
-            append_to_request(mutation, yyyymmdd, "quotes", quotes);
-	    if (mutation.entries_size() >= batch_size) {
-	      mutate_with_retries(*bt_stub, mutation);
-	    }
-          }
-          quotes.set_ticker(q.ticker());
-          quotes.clear_timestamp_ns();
-          quotes.clear_bid_px();
-          quotes.clear_bid_qty();
-          quotes.clear_offer_px();
-          quotes.clear_offer_qty();
-        }
-        quotes.add_timestamp_ns(q.timestamp_ns());
-        quotes.add_bid_px(q.bid_px());
-        quotes.add_bid_qty(q.bid_qty());
-        quotes.add_offer_px(q.offer_px());
-        quotes.add_offer_qty(q.offer_qty());
-
-        if (count % report_progress_rate == 0) {
-          auto elapsed = steady_clock::now() - upload_start;
-          std::cout << count << " quotes in " << chunk_count << " chunks.  elapsed-time="
-                    << duration_cast<seconds>(elapsed).count() << "s" << std::endl;
-        }
-      }
-    }
-  }
-  // ... CS101: the last batch needs to be uploaded too ...
-  append_to_request(mutation, yyyymmdd, "quotes", quotes);
-  mutate_with_retries(*bt_stub, mutation);
-
-  auto elapsed = steady_clock::now() - upload_start;
-  std::cout << count << " quotes in " << chunk_count << " chunks.  elapsed-time="
-	    << duration_cast<seconds>(elapsed).count() << "s" << std::endl;
-}
 
 void collate_trades(std::shared_ptr<grpc::ChannelCredentials> credentials,
                     std::string const& table_prefix,
